Test API and Service

+++ TheAbstraction.Application/Commands/Product/Delete/DeleteProductCommand.cs
@@ -1,5 +1,6 @@
 using TheAbstraction.Application.Common.Interfaces;
 using MediatR;
+using TheAbstraction.Application.Common.Exceptions;
 
 namespace TheAbstraction.Application.Commands.Product.Delete
 {
@@ -14,6 +15,8 @@ namespace TheAbstraction.Application.Commands.Product.Delete
 
         public Task<int> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
+            var productId = _productService.GetByIdAsync(request.Id).Result.Id;
+            if (productId == null) { throw new NotFoundException("Không tìm thấy người dùng"); }
             return _productService.DeleteAsync(request.Id, cancellationToken);
         }
     }
